Add first unit test
