+++ src/azure-blob/index.js
@@ -4,6 +4,7 @@ const multer = require('multer');
 const fs = require('fs');
 const path = require('path');
 require('dotenv').config();
+const handlers = require('./handlers');
 
 const { BlobServiceClient, StorageSharedKeyCredential } = require('@azure/storage-blob');
 
@@ -27,76 +28,11 @@ const containerClient = blobServiceClient.getContainerClient(process.env.AZURE_C
 app.use(express.static(path.join(__dirname, 'public')));
 app.use(express.json());
 
-app.post('/upload', upload.single('file'), async (req, res) => {
-    const fileName = req.body.note;
-    if (!fileName) {
-        return res.status(400).send('File name is required.');
-    }
+app.post('/upload', upload.single('file'), handlers.uploadFile);
 
-    if (req.file) {
-        try {
-            const connection = await mysql.createConnection({
-                host: process.env.DATABASE_HOST,
-                user: process.env.DATABASE_USER,
-                port: process.env.DATABASE_PORT,
-                database: process.env.DATABASE_NAME,
-            });
+app.get('/files', handlers.getFiles);
 
-            const blobName = req.file.filename;
-            const blockBlobClient = containerClient.getBlockBlobClient(blobName);
-
-            await blockBlobClient.uploadFile(req.file.path);
-            fs.unlinkSync(req.file.path); // remove the file locally after upload
-
-            connection.query('INSERT INTO file (name, filekey) VALUES (?,?)', [fileName, blobName]);
-            connection.end();
-
-            res.status(200).send('File uploaded successfully.');
-        } catch (err) {
-            console.error('Error uploading file:', err);
-            res.status(500).send('Failed to upload file.');
-        }
-    } else {
-        res.status(400).send('No file uploaded.');
-    }
-});
-
-app.get('/files', async(req, res) => {
-    const connection = await mysql.createConnection({
-        host: process.env.DATABASE_HOST,
-        user: process.env.DATABASE_USER,
-        port: process.env.DATABASE_PORT,
-        database: process.env.DATABASE_NAME,
-    });
-
-    const results  = await connection.query('SELECT * FROM file');
-    res.json(results[0]);
-    connection.end();
-});
-
-app.delete('/files/:key', async (req, res) => {
-    const fileKey = req.params.key;
-
-    try {
-        const connection = await mysql.createConnection({
-            host: process.env.DATABASE_HOST,
-            user: process.env.DATABASE_USER,
-            port: process.env.DATABASE_PORT,
-            database: process.env.DATABASE_NAME,
-        });
-
-        const blockBlobClient = containerClient.getBlockBlobClient(fileKey);
-        await blockBlobClient.delete();
-
-        connection.query('DELETE FROM file WHERE fileKey = ?', [fileKey]);
-        connection.end();
-
-        res.status(200).send('File deleted successfully.');
-    } catch (err) {
-        console.error('Error deleting file:', err);
-        res.status(500).send('Failed to delete file.');
-    }
-});
+app.delete('/files/:key', handlers.deleteFile);
 
 app.listen(PORT, () => {
     console.log(`Server is running on port ${PORT}`);
