+++ src/azure-blob/package.json
@@ -17,7 +17,7 @@
     "express": "^4.17.1",
     "multer": "^1.4.2",
     "mysql2": "^3.10.2",
-    "jest": "^26.6.3"
+    "jest": "^27.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^8.0.0",
